+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **aiken-lsp**: hover and goto definition support on list tail. @rvcas
 - **aiken-lsp**: hover on prop test via expression. @rvcas
 - **aiken-lang**: a new way to emit logs that don't get erased. @micahkendall
+- **aiken-lang**: new builtin types in the prelude `Pair` and `Pairs`. @KtorZ @Microproofs
 
 ### Fixed